requirements.txt
streamlit
pandas>=1.3.0
seaborn>=0.11.0
matplotlib>=3.4.0
langchain>=0.0.134
langchain-google-genai>=0.0.4
python-dotenv
